Add .master() call when creating Spark session
This is a new addition to pyspark since I last worked on this
code.

Note: the "local" argument appears to only be relevant for
standalone mode -- we should allow users to specify a master
name instead (e.g. "yarn" or "mesos" depending on how their
cluster is managed).

What we have now is fine for testing and development right now,
but will need to be fixed before this code can be used on a
proper Spark cluster...